requirements.txt
streamlit>=1.28.0
gspread>=5.12.0
google-auth>=2.23.0
toml>=0.10.2
nltk>=3.8.0
